Add login page to overall logic
