--- routes/login.js
@@ -1,8 +1,0 @@
-
-/*
- * GET login page.
- */
-
-exports.view = function(req, res){
-    res.render('login');
-};
