+++ views/add.handlebars
@@ -28,7 +28,7 @@
                         <li><a href='/'>Home</a></li>
                         <li><a href='/add'>Add Activity</a></li>
                         <li><a href='/report'>Report</a></li>
-                        <li><a href='/login'>Log Out</a></li>
+                        <li><a href='/'>Log Out</a></li>
                     </ul>
                 </div>
 
@@ -67,6 +67,16 @@
     </body>
 
     <script src='js/navbar.js'></script>
+    <script src='js/cookies.js'></script>
+	<script>
+		var username = getCookie('username', document.cookie);
+		if(username === '') {
+			// Not logged in
+			window.location.href = '/';
+		} else {
+			document.getElementById('username').innerText = username;
+		}
+	</script>
 
     <script type="text/javascript">
         $(document).ready(function(){
